Made the rc4 scripts as two python files and added bash file for running
